# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barangay Santiago Admin Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalResidents: number;
  activeResidents: number;
  pendingResidents: number;
  totalBlotterReports: number;
  pendingBlotterReports: number;
  investigatingBlotterReports: number;
  totalProjects: number;
  ongoingProjects: number;
  completedProjects: number;
  totalAnnouncements: number;
  publishedAnnouncements: number;
  totalOrdinances: number;
  enactedOrdinances: number;
  totalDocumentRequests: number;
  pendingDocumentRequests: number;
  totalAssets: number;
}

export interface ActivityItem {
  id: string;
  kind: string;
  title: string;
  description: string;
  timestamp: string;
}

export interface Resident {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  purok: string;
  gender: string;
  civilStatus: string;
  birthDate: string;
  address: string;
  documentType?: string;
  status: string;
  createdAt: string;
}

export interface ResidentInput {
  fullName: string;
  email: string;
  phone: string;
  purok: string;
  gender: string;
  civilStatus: string;
  birthDate: string;
  address: string;
  documentType?: string;
  status?: string;
}

export type ResidentStatsByPurokItem = {
  purok: string;
  count: number;
};

export interface ResidentStats {
  total: number;
  active: number;
  pending: number;
  thisMonth: number;
  byPurok: ResidentStatsByPurokItem[];
}

export interface BlotterReport {
  id: number;
  referenceNo: string;
  reporter: string;
  category: string;
  location: string;
  dateReported: string;
  status: string;
  description: string;
  resolutionNotes?: string;
  createdAt: string;
}

export interface BlotterReportInput {
  reporter: string;
  category: string;
  location: string;
  dateReported: string;
  status?: string;
  description: string;
  resolutionNotes?: string;
}

export interface BlotterStats {
  total: number;
  pending: number;
  investigating: number;
  resolved: number;
}

export interface Project {
  id: number;
  title: string;
  category: string;
  description: string;
  budget: number;
  startDate: string;
  targetDate: string;
  projectLeader: string;
  status: string;
  progress: number;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  category: string;
  description: string;
  budget: number;
  startDate: string;
  targetDate: string;
  projectLeader: string;
  status?: string;
  progress?: number;
}

export interface ProjectStats {
  total: number;
  ongoing: number;
  planning: number;
  completed: number;
  totalBudget: number;
}

export interface Announcement {
  id: number;
  type: string;
  title: string;
  content: string;
  eventDate?: string;
  eventTime?: string;
  location?: string;
  status: string;
  createdAt: string;
}

export interface AnnouncementInput {
  type: string;
  title: string;
  content: string;
  eventDate?: string;
  eventTime?: string;
  location?: string;
  status?: string;
}

export interface AnnouncementStats {
  total: number;
  published: number;
  draft: number;
  events: number;
}

export interface Ordinance {
  id: number;
  ordinanceNumber: string;
  title: string;
  description: string;
  ordinanceType: string;
  author: string;
  status: string;
  enactedDate?: string;
  createdAt: string;
}

export interface OrdinanceInput {
  ordinanceNumber: string;
  title: string;
  description: string;
  ordinanceType: string;
  author: string;
  status?: string;
  enactedDate?: string;
}

export interface OrdinanceStats {
  total: number;
  enacted: number;
  draft: number;
  repealed: number;
}

export interface DocumentRequest {
  id: number;
  referenceNo: string;
  residentName: string;
  documentType: string;
  purpose: string;
  price: number;
  status: string;
  paymentMethod?: string;
  requestedDate: string;
  createdAt: string;
}

export interface DocumentRequestInput {
  residentName: string;
  documentType: string;
  purpose: string;
  price: number;
  status?: string;
  paymentMethod?: string;
  requestedDate?: string;
}

export interface DocumentCategory {
  id: number;
  name: string;
  category: string;
  description: string;
  price: number;
  available: boolean;
}

export interface DocumentStats {
  total: number;
  pending: number;
  processing: number;
  ready: number;
  claimed: number;
}

export interface Asset {
  id: number;
  fileName: string;
  type: string;
  sizeKb: number;
  category: string;
  description?: string;
  uploadedBy: string;
  uploadedDate: string;
  createdAt: string;
}

export interface AssetInput {
  fileName: string;
  type: string;
  sizeKb: number;
  category: string;
  description?: string;
  uploadedBy: string;
  uploadedDate?: string;
}

export interface AssetStats {
  total: number;
  documents: number;
  images: number;
  videos: number;
}

export type ListResidentsParams = {
  search?: string;
  purok?: string;
  status?: ListResidentsStatus;
};

export type ListResidentsStatus =
  (typeof ListResidentsStatus)[keyof typeof ListResidentsStatus];

export const ListResidentsStatus = {
  active: "active",
  pending: "pending",
  inactive: "inactive",
} as const;

export type ListBlotterReportsParams = {
  search?: string;
  status?: string;
  category?: string;
};

export type ListProjectsParams = {
  search?: string;
  status?: string;
};

export type ListAnnouncementsParams = {
  search?: string;
  type?: string;
  status?: string;
};

export type ListOrdinancesParams = {
  search?: string;
  status?: string;
};

export type ListDocumentRequestsParams = {
  search?: string;
  status?: string;
  type?: string;
};

export type ListAssetsParams = {
  search?: string;
  type?: string;
};