# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barangay Santiago Admin Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blotter:
    get:
      operationId: listBlotterReports
      tags: [blotter]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlotterReport"
    post:
      operationId: createBlotterReport
      tags: [blotter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlotterReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlotterReport"
...
  /blotter/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getBlotterReport
      tags: [blotter]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlotterReport"
    patch:
      operationId: updateBlotterReport
      tags: [blotter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlotterReportInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlotterReport"
    delete:
      operationId: deleteBlotterReport
      tags: [blotter]
      responses:
        "204":
          description: Deleted
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
...
  /announcements/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getAnnouncement
      tags: [announcements]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      responses:
        "204":
          description: Deleted
...
  /documents:
    get:
      operationId: listDocumentRequests
      tags: [documents]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentRequest"
    post:
      operationId: createDocumentRequest
      tags: [documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRequest"
...
  /documents/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getDocumentRequest
      tags: [documents]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRequest"
    patch:
      operationId: updateDocumentRequest
      tags: [documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentRequestInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRequest"
    delete:
      operationId: deleteDocumentRequest
      tags: [documents]
      responses:
        "204":
          description: Deleted
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
...
  /assets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getAsset
      tags: [assets]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    patch:
      operationId: updateAsset
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    BlotterReport:
      type: object
      properties:
        id: { type: integer }
        referenceNo: { type: string }
        reporter: { type: string }
        category: { type: string }
        location: { type: string }
        dateReported: { type: string }
        status: { type: string }
        description: { type: string }
        resolutionNotes: { type: string }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - referenceNo
        - reporter
        - category
        - location
        - dateReported
        - status
        - description
        - createdAt
    BlotterReportInput:
      type: object
      properties:
        reporter: { type: string }
        category: { type: string }
        location: { type: string }
        dateReported: { type: string }
        status: { type: string }
        description: { type: string }
        resolutionNotes: { type: string }
      required:
        - reporter
        - category
        - location
        - dateReported
        - description
...
    Announcement:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        title: { type: string }
        content: { type: string }
        eventDate: { type: string }
        eventTime: { type: string }
        location: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - type
        - title
        - content
        - status
        - createdAt
    AnnouncementInput:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        content: { type: string }
        eventDate: { type: string }
        eventTime: { type: string }
        location: { type: string }
        status: { type: string }
      required: [type, title, content]
...
    DocumentRequest:
      type: object
      properties:
        id: { type: integer }
        referenceNo: { type: string }
        residentName: { type: string }
        documentType: { type: string }
        purpose: { type: string }
        price: { type: number }
        status: { type: string }
        paymentMethod: { type: string }
        requestedDate: { type: string }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - referenceNo
        - residentName
        - documentType
        - purpose
        - price
        - status
        - requestedDate
        - createdAt
    DocumentRequestInput:
      type: object
      properties:
        residentName: { type: string }
        documentType: { type: string }
        purpose: { type: string }
        price: { type: number }
        status: { type: string }
        paymentMethod: { type: string }
        requestedDate: { type: string }
      required: [residentName, documentType, purpose, price]
...
    Asset:
      type: object
      properties:
        id: { type: integer }
        fileName: { type: string }
        type: { type: string }
        sizeKb: { type: integer }
        category: { type: string }
        description: { type: string }
        uploadedBy: { type: string }
        uploadedDate: { type: string }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - fileName
        - type
        - sizeKb
        - category
        - uploadedBy
        - uploadedDate
        - createdAt
    AssetInput:
      type: object
      properties:
        fileName: { type: string }
        type: { type: string }
        sizeKb: { type: integer }
        category: { type: string }
        description: { type: string }
        uploadedBy: { type: string }
        uploadedDate: { type: string }
      required: [fileName, type, sizeKb, category, uploadedBy]
...