# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barangay Santiago Admin Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  fileName: string;
  type: string;
  sizeKb: number;
  category: string;
  description?: string;
  uploadedBy: string;
  uploadedDate: string;
  createdAt: Date;
}